Update combine_PRS.R
I came across PGS files with a alternate column name "#IID" instead of "IID". Instead of manually changing the column name for every PGS file it will now recognize altered names. The only prerequisite is that the IID column's alternate name contains the pattern "IID". The name is subsequently change to "IID" only if there is not a complete pattern match to allow for IID merging later in the script.