Update 2 combine_PRS.R
The current script cannot handle multiple PGS score files that share the same name for SUM column. Created a new option to specify the new names for the columns. I also altered the sub-setting of trait specific PGS to reflect the change in names. Added an error statement to indicate whether the new name list does not have the same length as the number of PGS scores under consideration.